Refactored the Entire program into different kotlin files and created the child TradeAPI class. Furthermore, i've expanded on the functionality of what you can query through the API. All I have to do from this point is add the news API into the program
