+++ app/src/main/java/com/example/totallylegal/HomeScreen.kt
@@ -1,0 +1,44 @@
+package com.example.totallylegal
+
+import android.util.Log
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import kotlinx.coroutines.launch
+
+@Composable
+fun HomeScreen() {
+    val tradeList = remember { mutableStateOf<List<Map<String, Any>>>(emptyList()) }
+    val coroutineScope = rememberCoroutineScope()
+
+    // Fetch API data when the screen is loaded
+    LaunchedEffect(Unit) {
+        coroutineScope.launch {
+            tradeList.value = TradeAPI().fetchByRepresentative("Nancy Pelosi")
+            Log.d("Size", tradeList.value.size.toString())
+        }
+    }
+
+    LazyColumn(
+        modifier = Modifier.fillMaxSize().padding(16.dp),
+        verticalArrangement = Arrangement.spacedBy(10.dp)
+    ) {
+        items(tradeList.value.size) { tradeItem ->
+            val ref = tradeList.value.get(tradeItem)
+            TradeBox(
+                ref.getValue("representative").toString(),
+                ref.getValue("ticker").toString(),
+                ref.getValue("amount").toString(),
+                ref.getValue("type").toString()
+            )
+        }
+    }
+}
